added translationHelps container
