+++ src/Container.js
@@ -8,73 +8,29 @@ import {
   getContextId,
   getManifest,
   getProjectSaveLocation,
-  getCurrentToolName
+  getCurrentToolName,
+  getCurrentProjectToolsSelectedGL
 } from './selectors';
 // helpers
 import * as settingsHelper from './helpers/settingsHelper';
-import * as tHelpsHelpers from './helpers/tHelpsHelpers';
 //containers
 import GroupMenuContainer from './containers/GroupMenuContainer';
 import VerseCheckContainer from './containers/VerseCheckContainer';
+import TranslationHelpsContainer from './containers/TranslationHelpsContainer';
 
 //ui-kit components
-import {CheckInfoCard, TranslationHelps} from 'tc-ui-toolkit';
+import {CheckInfoCard} from 'tc-ui-toolkit';
 
 class Container extends React.Component {
   constructor(props) {
     super(props);
+    this.state = {
+      showHelps: false
+    };
     this.toggleHelps = this.toggleHelps.bind(this);
-    this.toggleHelpsModal = this.toggleHelpsModal.bind(this);
-    this.followTHelpsLink = this.followTHelpsLink.bind(this);
   }
-
   componentWillMount() {
-    let selections = Array.from(this.props.selectionsReducer.selections);
-    this.setState({selections});
     settingsHelper.loadCorrectPaneSettings(this.props, this.props.actions.setToolSettings);
-    this._reloadArticle(this.props);
-  }
-
-  componentWillReceiveProps(nextProps) {
-    const {contextIdReducer, resourcesReducer} = this.props || {};
-    const nextContextIDReducer = nextProps.contextIdReducer;
-    if (contextIdReducer !== nextContextIDReducer) {
-      this._reloadArticle(nextProps);
-    }
-
-    const {contextId} = contextIdReducer;
-    const nextContextId = nextContextIDReducer.contextId;
-
-    const currentArticle = tHelpsHelpers.getArticleFromState(resourcesReducer, contextId);
-    const nextArticle = tHelpsHelpers.getArticleFromState(nextProps.resourcesReducer, nextContextId);
-    if (currentArticle !== nextArticle) {
-      var page = document.getElementById("helpsbody");
-      if (page) page.scrollTop = 0;
-    }
-  }
-
-  /**
-   * Loads the resource article
-   * @param props
-   * @private
-   */
-  _reloadArticle(props) {
-    const {contextIdReducer, toolsReducer, projectDetailsReducer, actions} = props;
-    const {contextId} = contextIdReducer;
-    if (contextId) {
-      const articleId = contextId.groupId;
-      const {currentToolName} = toolsReducer;
-      const languageId = projectDetailsReducer.currentProjectToolsSelectedGL[currentToolName];
-      actions.loadResourceArticle(currentToolName, articleId, languageId);
-    }
-  }
-
-  toggleHelps() {
-    this.setState({showHelps: !this.state.showHelps});
-  }
-
-  toggleHelpsModal() {
-    this.setState({showHelpsModal: !this.state.showHelpsModal});
   }
 
   getCheckInfoCardText(translationWords, articleId, translationHelps) {
@@ -106,35 +62,8 @@ class Container extends React.Component {
     return finalString;
   }
 
-  followTHelpsLink(link) {
-    let linkParts = link.split('/'); // link format: <lang>/<resource>/<category>/<article>
-
-    const [lang, type, category, article] = linkParts;
-    const resourceDir = tHelpsHelpers.getResourceDirByType(type);
-
-    this.props.actions.loadResourceArticle(resourceDir, article, lang, category);
-    const articleData = this.props.resourcesReducer.translationHelps[resourceDir][article];
-
-    let newState;
-    const tHelpsModalVisibility = true;
-    const articleCategory = category;
-    if (articleData) {
-      newState = {
-        tHelpsModalVisibility,
-        articleCategory,
-        modalArticle: articleData
-      };
-    } else {
-      newState = {
-        tHelpsModalVisibility,
-        articleCategory,
-        modalArticle: 'Cannot find an article for ' + link
-      };
-    }
-    //todo: Shouldn't need to to set state and return state in the same function
-    // Seems like an anti pattern
-    this.setState(newState);
-    return newState;
+  toggleHelps() {
+    this.setState({showHelps: !this.state.showHelps});
   }
 
   render() {
@@ -154,19 +83,13 @@ class Container extends React.Component {
       loginReducer,
       actions,
     } = this.props;
-    window.followLink = this.followTHelpsLink;
 
     if (contextId !== null) {
       let {translationWords} = resourcesReducer.translationHelps ? resourcesReducer.translationHelps : {};
-      const languageId = currentProjectToolsSelectedGL[currentToolName];
       const {groupId} = contextId;
       const title = groupsIndexReducer.groupsIndex.filter(item => item.id === groupId)[0].name;
-      // const glQuote = actions.getGLQuote(languageId, groupId, currentToolName);
       const checkInfoCardPhrase = this.getCheckInfoCardText(translationWords, contextId.groupId, resourcesReducer.translationHelps);
-      //get tHelps article
-      const currentFile = tHelpsHelpers.getArticleFromState(resourcesReducer, contextId);
-      const currentFileMarkdown = tHelpsHelpers.convertMarkdownLinks(currentFile, languageId);
-      const tHelpsModalMarkdown = tHelpsHelpers.convertMarkdownLinks(this.state.modalArticle, languageId, this.state.articleCategory);
+      // const glQuote = actions.getGLQuote(languageId, groupId, currentToolName);
       return (
         <div style={{display: 'flex', flexDirection: 'row'}}>
           <GroupMenuContainer {...this.props.groupMenu} />
@@ -179,13 +102,10 @@ class Container extends React.Component {
               onSeeMoreClick={this.toggleHelps.bind(this)} />
             <VerseCheckContainer {...this.props.verseCheck} />
           </div>
-          <TranslationHelps
-            article={currentFileMarkdown}
-            modalArticle={tHelpsModalMarkdown}
-            openExpandedHelpsModal={() => this.toggleHelpsModal()}
-            isShowHelpsSidebar={this.state.showHelps}
-            sidebarToggle={() => this.toggleHelps()}
-            isShowHelpsExpanded={this.state.showHelpsModal} />
+          <TranslationHelpsContainer
+            toggleHelps={this.toggleHelps.bind(this)}
+            showHelps={this.state.showHelps}
+            {...this.props.translationHelps} />
         </div>
       );
     } else {
@@ -249,6 +169,13 @@ const mapStateToProps = (state, ownProps) => {
       toolsReducer: ownProps.tc.toolsReducer,
       groupsDataReducer: ownProps.tc.groupsDataReducer,
       remindersReducer: ownProps.tc.remindersReducer,
+      actions: ownProps.tc.actions
+    },
+    translationHelps: {
+      currentProjectToolsSelectedGL: getCurrentProjectToolsSelectedGL(ownProps),
+      toolsReducer: ownProps.tc.toolsReducer,
+      resourcesReducer: ownProps.tc.resourcesReducer,
+      contextIdReducer: ownProps.tc.contextIdReducer,
       actions: ownProps.tc.actions
     }
   };
